chore(quality-gate): green the workspace clippy -D warnings + rustfmt gate
Clears accumulated debt across the session's merged code so
`cargo clippy --workspace --tests --examples -- -D warnings` and
`cargo fmt --all --check` pass (both CI-enforced). No behavior change.

- cast_possible_truncation/precision_loss: file-level allow on the new model
  files, matching the 250+ existing model files (llama.rs et al). All sites are
  grid/index/dimension/serialization casts.
- dead_code: scoped allows on deserialized-verbatim config fields, backend-gated
  helpers (vulkan/cuda/rocm), and accessors, matching the model-file convention.
- unused imports removed (pixal3d Module, function-local SpeculativeProposer).
- needless_range_loop/too_many_arguments: scoped allows (rewriting tensor loops
  or arg-lists would risk numerics; matches the 343 existing too_many_args allows).
- ring_exchange cfg-gated to feature=ring (its only callers are ring-gated).
- dropped the inert route_handler utoipa::path annotation (not in the OpenApi
  paths() registry; RouteResponse's Decision can't be ToSchema).
- misc idiom fixes (sort_by_key, checked_div, struct-update literal, doc lists).
- rustfmt the merged files (pixal3d, film, oasis, qwen3_omni, ...).

+++ hanzo-engine/src/diffusion_models/ace_step/pipeline.rs
@@ -107,7 +107,7 @@ mod tests {
         out.extend_from_slice(&1u16.to_le_bytes()); // PCM
         out.extend_from_slice(&ch.to_le_bytes());
         out.extend_from_slice(&sr.to_le_bytes());
-        out.extend_from_slice(&(sr * bytes_per as u32 * ch as u32).to_le_bytes());
+        out.extend_from_slice(&(sr * bytes_per * ch as u32).to_le_bytes());
         out.extend_from_slice(&(bytes_per as u16 * ch).to_le_bytes());
         out.extend_from_slice(&16u16.to_le_bytes());
         out.extend_from_slice(b"data");
